+++ IoT/IoT_embedded/src/calibration.h
@@ -1,0 +1,7 @@
+#ifndef CALIBRATION_H
+#define CALIBRATION_H
+
+void calibrate_water_levels(void);
+void calibrate_sprinkler_angles(void);
+
+#endif
