+++ IoT/IoT_embedded/src/display.h
@@ -1,0 +1,8 @@
+#ifndef DISPLAY_UTILS_H
+#define DISPLAY_UTILS_H
+
+void asyncDisableDisplayAfterMs(int ms);
+void displayAndDie(void);
+void init_display(void);
+
+#endif
