+++ IoT/IoT_embedded/src/leds.h
@@ -1,0 +1,7 @@
+#ifndef LEDS_UTILS_H
+#define LEDS_UTILS_H
+
+void turnOffAll(void);
+void ledAnimation(void);
+
+#endif
