+++ IoT/IoT_embedded/src/sensors.h
@@ -1,0 +1,8 @@
+#ifndef SENSORS_H
+#define SENSORS_H
+
+void init_sensors(void);
+void collect_all_sensor_data(void);
+extern char outbound_buffer[256];
+
+#endif
