+++ IoT/IoT_embedded/src/watering.h
@@ -1,0 +1,10 @@
+#ifndef WATERING_H
+#define WATERING_H
+
+void init_watering(void);
+void runWaterPump(void);
+void waterPlant1(void);
+void waterPlant2(void);
+void neutral(void);
+
+#endif
